requirements.txt
openai[datalib]==0.27.8
tiktoken==0.4.0
python-dotenv==1.0.0
python-slugify==8.0.1
